content_indexer v0.1.0 ContentIndexer.Services.Indexer

Summary

Indexer is a Genserver that holds the index state - basically a list of index structs that have the filename, tokens and weights
Each time an index struct is added to the server/index the weightings are re-calculated. Since they are stored in memory the index searching is fast

Functions

add(file_name, tokens)

calculate()

corpus_of_tokens()

documents()

init_indexer()

reset_index()

retrieve_index()

start_link()