content_indexer v0.1.0 ContentIndexer.Services.ListCheckerWorker

genserver based approach to the ListCheckerWorker Summary

ListCheckerWorker is the OTP actor that handles the actual ContentIndexerService.list_contains to check
whether a given word is contained in a list of tokens

Due to the fact that we have a very large dataset - potentially millions - a process is spawned for each

Basic Useage

A method is called by the server for each list of tokens and returns a result based on whether a given word was found in the list or not

Functions

init_worker()

list(index, word, tokens)

start_link()