content_indexer v0.2.3 ContentIndexer.TfIdf.Corpus

Summary Corpus is a Genserver that simply holds the total number of docs in the index

Functions

count()

Count of all documents in the index

## Example

iex> ContentIndexer.TfIdf.Corpus.count
{:ok, 23}

decrement()

Decrement the count of all documents in the index

## Example

iex> ContentIndexer.TfIdf.Corpus.decrement
{:ok, :decremented}

increment()

Increment the count of all documents in the index

## Example

iex> ContentIndexer.TfIdf.Corpus.increment
{:ok, :incremented}

init_corpus()

Initialises the Corpus doc count with zero

reset()

Resets the corpus document count

## Example

iex> ContentIndexer.TfIdf.Corpus.reset
{:ok, 0}

start_link()