content_indexer v0.2.4 ContentIndexer.TfIdf.DocCounts

Summary DocCounts is a GenServer that contains a map of documents with their respective total number of terms in the document

Functions

add_document(document_name, term_count)

If the document exists update the count, if it’s a new document add it and it’s respective term_count

## Example

iex> ContentIndexer.TfIdf.DocCounts.add_document("test_file_1.md", 23)
{:ok, {"test_file_1.md", 23}}

document_term_count(document_name)

Retrieves the current term_count for this document or nil if not in the map

## Example

iex> ContentIndexer.TfIdf.DocCounts.document_term_count("test_file_1.md")
{:ok, {"test_file_1.md", 23}}

init_doc_counts()

Initialises the Doc Counts with an empty map

reset()

Resets the document counts

## Example

iex> ContentIndexer.TfIdf.DocCounts.reset
{:ok, 0}

start_link()

state()

Retrieves the current set of doc_counts i.e. the state

## Example

iex> ContentIndexer.TfIdf.DocCounts.state
{:ok, 0}