content_indexer v0.2.5 ContentIndexer.TfIdf.DocCounts.Impl

Summary functions used by the ContentIndexer.TfIdf.DocCounts.Server

Functions

add_new_document(document_name, term_count, document_map)

If the document exists update the count, if it’s a new document add it and it’s respective term_count

## Example

iex> ContentIndexer.TfIdf.DocCounts.Impl.add_document("test_file_1.md", 23)
      {"test_file_1.md", 23}

find_term_count(document_name, document_map)

Retrieves the current term_count for this document or nil if not in the map

## Example

iex> ContentIndexer.TfIdf.DocCounts.Impl.document_term_count("test_file_1.md")
    {"test_file_1.md", 23}

reset()

Resets the document counts

## Example

iex> ContentIndexer.TfIdf.DocCounts.Impl.reset()
  %{}

state(state)

Returns the document counts state

## Example

iex> ContentIndexer.TfIdf.DocCounts.Impl.state()
  %{}