content_indexer v0.2.5 ContentIndexer.TfIdf.TermCounts.Impl

Summary functions used by the ContentIndexer.TfIdf.TermCounts.Server

Functions

get_term_count(term, terms)

If the term exists get the count else zero

## Example

iex> ContentIndexer.TfIdf.TermCounts.Impl.get_term_count("bread")
      3

increment_term_count(term, terms)

If the term exists increment the count, if it’s a new term add and set count to 1

## Example

iex> ContentIndexer.TfIdf.TermCounts.increment_term_count("bread")
      {"bread", 3}

reset()

Resets the term counts

## Example

iex> ContentIndexer.TfIdf.TermCounts.Impl.reset
      %{}

state(state)

Retrieves the current set of term_counts i.e. the state

## Example

iex> ContentIndexer.TfIdf.TermCounts.Impl.state(%{})
      %{}